cloud/oats_cloud_setup/inventory/host_vars/cloud_setup/files/location_data_processing/requirements.txt
joblib
numpy
pandas
psycopg2-binary
pykalman
pytz
scipy